pyproject.toml
[tool.poetry]
name = "shortify"
version = "0.1.7"
description = "Shortify is a URL shortener RESTful API built with Python and FastAPI."
authors = ["Iliya <IHosseini083@gmail.com>"]
license = "GPL-3.0-only"
readme = "README.md"
repository = "https://github.com/IHosseini/Shortify"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = {extras = ["all"], version = "^0.95.1"}
beanie = "^1.18.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
structlog = "^23.1.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
ruff = "^0.0.264"
mypy = "^1.2.0"
black = "^23.3.0"
pre-commit = "^3.2.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/IHosseini/Shortify/issues"

[tool.isort]
profile = "black"
skip_gitignore = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 88

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true

[tool.ruff]
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black.
    "B008",  # do not perform function calls in argument defaults.
    "B010",  # Do not call `setattr` with a constant attribute value.
]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".mypy_cache",
    ".ruff_cache",
    ".venv",
    "venv",
    "__pycache__",
]

[tool.ruff.isort]
known-third-party = ["fastapi", "pydantic", "starlette"]

[tool.ruff.per-file-ignores]
# Ignore `F401` violations in any `__init__.py` file.
"__init__.py" = ["F401"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
